+++ nix/hosts/dionysus/default.nix
@@ -17,6 +17,18 @@
 
   programs.steam = {
     enable = true;
+    extest.enable = true;
+    package = pkgs.steam.override {
+      extraEnv = {
+        STEAM_FORCE_WAYLAND = "1";
+        SDL_VIDEODRIVER = "wayland";
+        GDK_BACKEND = "wayland,x11";
+        QT_QPA_PLATFORM = "wayland;xcb";
+      };
+    };
+    extraPackages = with pkgs; [ gamescope ];
   };
+
+  programs.gamescope.enable = true;
 
 }